project/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-restframework>=0.0.1",
    "djangorestframework>=3.15.0",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.3",
    "jinja2>=3.1.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.0.0",
    "whitenoise>=6.11.0",
]
